+++ packages/use-intl/.eslintrc.js
@@ -1,3 +1,6 @@
 module.exports = {
-  extends: ['molindo/typescript', 'molindo/react']
+  extends: ['molindo/typescript', 'molindo/react'],
+  rules: {
+    'import/no-useless-path-segments': 'error'
+  }
 };